requirements.txt
numpy
pandas
scipy
scikit-learn
torch
torchvision
transformers
matplotlib
seaborn
Pillow
nibabel
jupyter
notebook
ipykernel
tqdm
pyyaml
plotly
vtk
opencv-python